Add basic Dockerfile
It creates a container that starts a virtual X server with Openbox, a VNC server
and then starts the Backblaze personal backup client via WINE.

You need to open up port 5900 and have a volume with a WINE prefix and already
installed Backblaze client mounted at /root/wine/ for this to work.
You can install it using this container too (If you can somehow get your
BZ credetials to go through the VNC properly that is).